+++ tests/state/test_maps.py
@@ -1,6 +1,6 @@
 import pytest
 
-from glassdolls.state.maps import MapState
+from glassdolls.state.map import MapState
 from glassdolls.constants import MAP_TOWN_TEST_FILE
